Dodanie modułu silnika zapisów
