+++ zapisy/apps/enrollment/records/engine/__init__.py
@@ -1,0 +1,3 @@
+from apps.enrollment.records.engine.fillup_group import fill_group
+
+__all__ = ['fill_group']
